# docs/swagger-ui/openapi.yaml
{
  "components": {
    "schemas": {
...
      "AgreementRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "RelatedPartyRef": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "validFor": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          }
        }
      },
      "AgreementTermOrCondition": {
        "type": "object",
        "properties": {
          "validFor": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "schemaLocation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
      "ProductRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "CredentialAttributeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "CredentialOfferDto": {
        "type": "object",
        "properties": {
          "credential_preview": {
            "$ref": "#/components/schemas/CredentialPreviewDto"
          }
        }
      },
      "PagedSlaResponse": {
        "type": "object",
        "properties": {
          "pagedSLAs": {
            "$ref": "#/components/schemas/PageServiceLevelAgreement"
          }
        }
      },
      "Sort": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "CredentialPreviewDto": {
        "type": "object",
        "properties": {
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialAttributeDto"
            }
          },
          "did": {
            "type": "string"
          }
        }
      },
...
      "AgreementSpecificationRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "TimePeriod": {
        "type": "object",
        "properties": {
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          }
        }
      },
...
      "Characteristic": {
        "type": "object",
        "properties": {
          "schemaLocation": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "AgreementAuthorization": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "signaturePresentation": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "ServiceLevelAgreement": {
        "type": "object",
        "properties": {
          "approved": {
            "type": "boolean"
          },
          "templateRef": {
            "$ref": "#/components/schemas/TemplateRef"
          },
          "approvalDate": {
            "format": "date-time",
            "type": "string"
          },
          "validFor": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            }
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "relatedParty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedPartyRef"
            }
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "RelatedParty": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "extendedInfo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "Rule": {
        "type": "object",
        "properties": {
          "unit": {
            "type": "string"
          },
          "consequence": {
            "type": "string"
          },
          "metric": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "tolerance": {
            "type": "string"
          },
          "referenceValue": {
            "type": "string"
          }
        }
      },
...
      "ProductOfferingRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
      "DIDStateDto": {
        "type": "object",
        "properties": {
          "updated_at": {
            "type": "string"
          },
          "credential_offer_dict": {
            "$ref": "#/components/schemas/CredentialOfferDto"
          },
          "schema_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "credential_exchange_id": {
            "type": "string"
          }
        }
      },
...
      "PageServiceLevelAgreement": {
        "type": "object",
        "properties": {
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "numberOfElements": {
            "format": "int32",
            "type": "integer"
          },
          "last": {
            "type": "boolean"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageable"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "first": {
            "type": "boolean"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceLevelAgreement"
            }
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "AgreementItem": {
        "type": "object",
        "properties": {
          "termOrCondition": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementTermOrCondition"
            }
          },
          "product": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductRef"
            }
          },
          "productOffering": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingRef"
            }
          },
          "schemaLocation": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "Agreement": {
        "type": "object",
        "properties": {
          "agreementPeriod": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "agreementType": {
            "type": "string"
          },
          "associatedAgreement": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementRef"
            }
          },
          "schemaLocation": {
            "type": "string"
          },
          "documentNumber": {
            "format": "int32",
            "type": "integer"
          },
          "agreementSpecification": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementSpecificationRef"
            }
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "initialDate": {
            "format": "date-time",
            "type": "string"
          },
          "characteristic": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Characteristic"
            }
          },
          "baseType": {
            "type": "string"
          },
          "agreementAuthorization": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementAuthorization"
            }
          },
          "name": {
            "type": "string"
          },
          "completionDate": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "statementOfIntent": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "agreementItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementItem"
            }
          },
          "engagedParty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedParty"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "TemplateRef": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "href": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8087"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/api/v1/agreement/{identifier}": {
      "get": {
        "description": "Get an agreement by id or DID",
        "operationId": "getAgreement",
        "responses": {
          "200": {
            "description": "An agreement matching the provided identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agreement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id was provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An  couldn\u0027t be found with the provided Identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "identifier",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
    "/api/v1/agreement/{id}/identity": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DIDStateDto"
              }
            }
          },
          "required": true
        },
        "description": "Callback endpoint to handle processing async DID identifier generation",
        "operationId": "updateTemplateIdentity_1",
        "responses": {
          "200": {
            "description": "The Agreement was updated successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An Agreement couldn\u0027t be found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
    "/api/v1/service-level-agreement/{identifier}": {
      "get": {
        "description": "Get an SLA by id or DID",
        "operationId": "getServiceLevelAgreement",
        "responses": {
          "200": {
            "description": "An SLA matching the provided identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceLevelAgreement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id was provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An SLA couldn\u0027t be found with the provided Identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "identifier",
            "required": true
          }
        ],
        "tags": [
          "Service Level Agreement"
        ]
      }
    },
    "/api/v1/service-level-agreement": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceLevelAgreement"
              }
            }
          },
          "required": true
        },
        "description": "Create a new Service Level Agreement definition",
        "operationId": "createServiceLevelAgreement",
        "responses": {
          "201": {
            "description": "A SLA definition was created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid SLA definition provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Service Level Agreement"
        ]
      },
      "get": {
        "description": "Retrieve a paged collection of Service Level Agreements according to paging and filter parameters",
        "operationId": "getServiceLevelAgreements",
        "responses": {
          "200": {
            "description": "A Paged List of Service Level Agreements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedSlaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid page or filter parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "default": "0",
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "description": "Page you want to retrieve (0..N)"
          },
          {
            "schema": {
              "default": "20",
              "type": "integer"
            },
            "in": "query",
            "name": "size",
            "description": "Number of records per page."
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "in": "query",
            "name": "sort",
            "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported."
          }
        ],
        "tags": [
          "Service Level Agreement"
        ]
      }
    },
    "/api/v1/agreement/{id}": {
      "delete": {
        "description": "Delete an Agreement definition by DID",
        "operationId": "removeAgreement",
        "responses": {
          "200": {
            "description": "The Agreement was deleted successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An Agreement couldn\u0027t be found with the provided DID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "did",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
    "/api/v1/service-level-agreement/{slaId}/identity": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DIDStateDto"
              }
            }
          },
          "required": true
        },
        "description": "Callback endpoint to handle processing async DID identifier generation",
        "operationId": "updateTemplateIdentity",
        "responses": {
          "200": {
            "description": "The SLA was updated successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An SLA couldn\u0027t be found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "slaId",
            "required": true
          }
        ],
        "tags": [
          "Service Level Agreement"
        ]
      }
    },
    "/api/v1/service-level-agreement/{did}": {
      "delete": {
        "description": "Delete an SLA definition by DID",
        "operationId": "removeServiceLevelAgreement",
        "responses": {
          "200": {
            "description": "The SLA was deleted successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An SLA couldn\u0027t be found with the provided DID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "did",
            "required": true
          }
        ],
        "tags": [
          "Service Level Agreement"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}